Refactor grid ribbon component
