+++ components/GridRibbon.tsx
@@ -3,7 +3,8 @@ import {
   Text
 } from '@chakra-ui/react'
 
-const ribbonStyled = (addOnStyle) => {
+// Reusable ribbon style accepting addon styling
+const ribbonStyle = (addOnStyle) => {
   return {
     content: '""',
     position: 'absolute',
@@ -16,14 +17,19 @@ const ribbonStyled = (addOnStyle) => {
   }
 }
 
-export default function GridRibbon({ribbonTitle}) {
+/** GridRibbon component is the UI to label insurance plans in 
+ ** the insurance catalog page or any page that contain any remark 
+ ** (i.e New, Trending, etc) in the grid. This component accepting 
+ ** string prop named as label
+**/
+export default function GridRibbon({label}) {
   // Default ribbon color - Purple
-  let mainRibbonColor = '#6758CA';
-  let backRibbonColor = '#52469E';
+  let mainRibbonColor = '#6758CA'
+  let backRibbonColor = '#52469E'
 
-  // Title 'Trending' set ribbon color - Red
-  if (ribbonTitle === 'Trending') {
-    mainRibbonColor = '#EB4d4B';
+  // Label 'Trending' set ribbon color - Red
+  if (label === 'Trending') {
+    mainRibbonColor = '#EB4d4B'
     backRibbonColor = '#D63031'
   }
 
@@ -35,8 +41,8 @@ export default function GridRibbon({ribbonTitle}) {
       position={'absolute'}
       top={'-10px'}
       right={'-10px'}
-      _after={ribbonStyled({bottom: 0, right: 0, border: `5px solid ${backRibbonColor}`})}
-      _before={ribbonStyled({top: 0, left: 0, border: `5px solid ${backRibbonColor}`})}
+      _after={ribbonStyle({bottom: 0, right: 0, border: `5px solid ${backRibbonColor}`})}
+      _before={ribbonStyle({top: 0, left: 0, border: `5px solid ${backRibbonColor}`})}
       >
       <Text
         as={'span'}
@@ -54,7 +60,7 @@ export default function GridRibbon({ribbonTitle}) {
         fontWeight={'bold'}
         textTransform={'uppercase'}
       >
-        {ribbonTitle}
+        {label}
       </Text>
     </Box>
   )
